refactor(test) : 당첨번호에 대한 검증을 WinningLotto로 책임 분리하기 위한 테스트 이관
