+++ src/main/java/lotto/model/WinningLotto.java
@@ -1,0 +1,8 @@
+package lotto.model;
+
+import java.util.List;
+
+public class WinningLotto {
+    public WinningLotto(List<Integer> numberList) {
+    }
+}
