+++ src/test/java/lotto/model/LottoPlayServiceTest.java
@@ -53,33 +53,7 @@ class LottoPlayServiceTest {
         assertEquals((double) 1_510_000 / 3000, lottos.getResultProfitRate());
     }
 
-    @DisplayName("지난주 로또 번호 숫자가 1~45 사이가 아닌 경우 검증")
-    @Test
-    void playLottoGame_not_lotto_number() {
-        assertThatIllegalArgumentException()
-                .isThrownBy(() -> lottoPlayService.playLottoGame(lottos, Arrays.asList(3, 7, 10, 35, 43, 46)))
-                .withMessage("[ERROR] 로또 번호는 1~45 사이의 숫자여야합니다.");
 
-        assertThatIllegalArgumentException()
-                .isThrownBy(() -> lottoPlayService.playLottoGame(lottos, Arrays.asList(0, 7, 10, 35, 43, 46)))
-                .withMessage("[ERROR] 로또 번호는 1~45 사이의 숫자여야합니다.");
-    }
-
-    @DisplayName("지난주 로또 번호 숫자들이 6개가 아닌 경우 검증")
-    @Test
-    void playLottoGame_non_six_number() {
-        assertThatIllegalArgumentException()
-                .isThrownBy(() -> lottoPlayService.playLottoGame(lottos, Arrays.asList(3, 7, 10, 35)))
-                .withMessage("[ERROR] 로또 번호는 6개여야 합니다.");
-    }
-
-    @DisplayName("지난주 로또 번호 6개의 숫자들에 중복이 있는지 검증")
-    @Test
-    void playLottoGame_duplication_number() {
-        assertThatIllegalArgumentException()
-                .isThrownBy(() -> lottoPlayService.playLottoGame(lottos, Arrays.asList(3, 7, 10, 10, 25, 35)))
-                .withMessage("[ERROR] 6개의 로또 번호에 중복이 있습니다.");
-    }
 
 
 }
